requirements.txt
opencv-python==4.5.1.48
numpy==1.20.2
matplotlib==3.4.1
torch==1.8.1
torchvision==0.9.1
torchinfo==0.0.9
tensorboard==2.4.1
Pillow==8.2.0
faiss-cpu==1.7.0
